+++ pkg/influx/macros_test.go
@@ -1,4 +1,4 @@
-package datasource
+package influx
 
 import (
 	"testing"
