+++ pkg/influx/mock.go
@@ -1,4 +1,4 @@
-package datasource
+package influx
 
 import (
 	"context"
