+++ pkg/influx/testdata/grouping.csv
