+++ pkg/influx/testdata/multiple.csv
