+++ pkg/influx/testdata/simple.csv
